Extract PipelineInventoryCalculateSQLBuilder from PipelineDataConsistencyCalculateSQLBuilder
